Revert "Merge remote-tracking branch 'origin/projectsInfoView_animation' into projectsInfoView_animation"
This reverts commit 19596c7374b8d48744896c49eb57340ee9955c59, reversing
changes made to 788ed4610d7fc0a887e9e5187cf89bdcd71744c1.
